Add done toast

+++ app/src/main/java/com/example/habit/data/model/HabitUID.kt
@@ -2,4 +2,4 @@ package com.example.habit.data.model
 
 data class HabitUID(val uid: String)
 
-data class HabitDone(val date: Long, val habit_uid: String)
+data class HabitDone(val date: Int, val habit_uid: String)
